requirements.txt
# Telegram Bot
python-telegram-bot[job-queue]==21.7

# HTTP Client
httpx==0.27.2

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.14.0
aiosqlite==0.20.0

# Configuration
pydantic-settings==2.6.1
python-dotenv==1.0.1

# RSS/Feed Parsing
feedparser==6.0.11

# Web Scraping (for Nature journals)
beautifulsoup4==4.12.3
lxml==5.3.0

# PDF/Export
markdown>=3.5.0
weasyprint>=60.0
aiofiles>=24.1.0

# Scheduling
apscheduler==3.10.4

# Date parsing
python-dateutil==2.9.0
